+++ R/udpipe_train.R
@@ -154,8 +154,9 @@ udpipe_train <- function(file = file.path(getwd(), "my_annotator.udpipe"),
 #' \url{http://universaldependencies.org/format.html}
 #' @export
 #' @examples 
-#' x <- udpipe_download_model(language = "dutch-lassysmall")
-#' ud_dutch <- udpipe_load_model(x$file_model)
+#' model <- udpipe_download_model(language = "dutch-lassysmall")
+#' if(!model$download_failed){
+#' ud_dutch <- udpipe_load_model(model$file_model)
 #' 
 #' file_conllu <- system.file(package = "udpipe", "dummydata", "traindata.conllu")
 #' metrics <- udpipe_accuracy(ud_dutch, file_conllu)
@@ -169,9 +170,11 @@ udpipe_train <- function(file = file.path(getwd(), "my_annotator.udpipe"),
 #' metrics <- udpipe_accuracy(ud_dutch, file_conllu, 
 #'                            tokenizer = "default", tagger = "none", parser = "none")
 #' metrics$accuracy
+#' }
+#' 
 #' 
 #' ## cleanup for CRAN only - you probably want to keep your model if you have downloaded it
-#' file.remove("dutch-lassysmall-ud-2.0-170801.udpipe")
+#' if(file.exists(model$file_model)) file.remove(model$file_model)
 udpipe_accuracy <- function(object, 
                             file_conllu, 
                             tokenizer = c("default", "none"),